requirements.txt
Django>=4.0.6,<4.1
djangorestframework>=3.13.1,<3.14
flake8>=5.0.4,<5.1
psycopg2>=2.8.6,<2.9